Capstone Project/Chirayu/RAG-chatapp/requirements.txt
python-dotenv~=1.0.1
torch~=2.2.2
transformers~=4.39.3
bitsandbytes==0.43.1
accelerate==0.28.0
pypdf~=4.1.0
tiktoken~=0.6.0
sentence-transformers~=2.5.1
faiss-cpu~=1.8.0
langchain~=0.1.16
streamlit~=1.29.0